requirements.txt
Flask==1.0.2
google-cloud-storage==1.16.1
numpy==1.16.2
requests==2.22.0